refactor(mcp-server): extract tool schemas and validation into modules
First slice of decomposing the mcp-server/index.ts god file. Moves two
self-contained, behavior-preserving groups out of index.ts:

- tool-validation.ts: VALID_SOURCES/TYPES, MAX_MEMORY_CONTENT_LENGTH,
  ADMIN_TOOLS_ENV, isNonEmptyText, validateMemoryContent,
  validateForgetSelection, admin-tool guards, initializeEmbeddingProvider.
- tool-schemas.ts: the nine memory_* zod tool-input schemas.

index.ts re-exports the public surface so tests and embedders importing from
`mcp-server/index.js` keep resolving unchanged. No behavior change: typecheck,
lint, format, and the full suite (785 passed, mcp-server 97/97) stay green.
